+++ crates/emmylua_code_analysis/src/compilation/analyzer/doc/type_ref_tags.rs
@@ -9,7 +9,8 @@ use crate::{
     db_index::{
         LuaDeclId, LuaDocParamInfo, LuaDocReturnInfo, LuaMemberId, LuaOperator, LuaPropertyOwnerId,
         LuaSignatureId, LuaType,
-    }, InFiled, LuaFlowId, SignatureReturnStatus, TypeAssertion
+    },
+    InFiled, LuaFlowId, SignatureReturnStatus, TypeAssertion,
 };
 
 use super::{
@@ -52,7 +53,7 @@ pub fn analyze_type(analyzer: &mut DocAnalyzer, tag: LuaDocTagType) -> Option<()
                             .db
                             .get_member_index_mut()
                             .get_member_mut(&member_id)?;
-                        member.decl_type = type_ref.clone();
+                        member.set_decl_type(type_ref.clone());
                     }
                 }
             }
@@ -79,7 +80,7 @@ pub fn analyze_type(analyzer: &mut DocAnalyzer, tag: LuaDocTagType) -> Option<()
                     .db
                     .get_member_index_mut()
                     .get_member_mut(&member_id)?;
-                member.decl_type = first_type.clone();
+                member.set_decl_type(first_type.clone());
             }
         }
         _ => {}