+++ crates/emmylua_code_analysis/src/compilation/test/flow.rs
@@ -100,7 +100,7 @@ mod test {
         "#
         ));
     }
-    
+
     #[test]
     fn test_issue_100() {
         let mut ws = VirtualWorkspace::new_with_init_std_lib();
@@ -141,6 +141,42 @@ mod test {
 
         foo(text)
         "#
+        ));
+    }
+
+    #[test]
+    fn test_null_function_field() {
+        let mut ws = VirtualWorkspace::new_with_init_std_lib();
+        assert!(ws.check_code_for(
+            DiagnosticCode::NeedCheckNil,
+            r#"
+        ---@class A
+        ---@field aaa? fun(a: string)
+
+
+        local c ---@type A
+
+        if c.aaa then
+            c.aaa("aaa")
+        end
+        "#
+        ))
+    }
+
+    #[test]
+    fn test_issue_162() {
+        let mut ws = VirtualWorkspace::new();
+        assert!(ws.check_code_for(
+            DiagnosticCode::ParamTypeNotMatch,
+            r#"
+            --- @class Foo
+            --- @field a? fun()
+
+            --- @param _o Foo
+            function bar(_o) end
+
+            bar({})
+            "#
         ));
     }
 }