+++ crates/emmylua_code_analysis/src/compilation/test/mod.rs
@@ -1,6 +1,9 @@
 mod closure_param_infer_test;
 mod closure_return_test;
 mod flow;
+mod inherit_type;
+mod mathlib_test;
 mod multi_return;
 mod out_of_order;
 mod overload_field;
+mod overload_test;