+++ crates/emmylua_code_analysis/src/db_index/member/mod.rs
@@ -26,31 +26,34 @@ impl LuaMemberIndex {
     pub fn add_member(&mut self, member: LuaMember) -> LuaMemberId {
         let id = member.get_id();
         let owner = member.get_owner();
-        let key = member.get_key().clone();
-        if !owner.is_none() {
-            self.owner_members
-                .entry(owner)
-                .or_insert_with(HashMap::new)
-                .entry(key).or_insert(id);
-        }
         let file_id = member.get_file_id();
         self.in_field_members
             .entry(file_id)
             .or_insert_with(Vec::new)
             .push(id);
         self.members.insert(id, member);
+
+        if !owner.is_none() {
+            self.add_member_owner(owner.clone(), id);
+            self.add_member_to_owner(owner, id);
+        }
         id
+    }
+
+    pub fn add_member_to_owner(&mut self, owner: LuaMemberOwner, id: LuaMemberId) -> Option<()> {
+        let member = self.members.get(&id)?;
+        let key = member.get_key().clone();
+        let member_map = self.owner_members.entry(owner).or_insert_with(HashMap::new);
+        if !member_map.contains_key(&key) {
+            member_map.insert(key, id);
+        }
+
+        Some(())
     }
 
     pub fn add_member_owner(&mut self, owner: LuaMemberOwner, id: LuaMemberId) -> Option<()> {
         let member = self.members.get_mut(&id)?;
-        let key = member.get_key().clone();
         member.owner = owner.clone();
-
-        self.owner_members
-            .entry(owner)
-            .or_insert_with(HashMap::new)
-            .entry(key).or_insert(id);
 
         Some(())
     }
@@ -81,7 +84,6 @@ impl LuaMemberIndex {
             None
         }
     }
-
 }
 
 impl LuaIndex for LuaMemberIndex {