+++ crates/emmylua_code_analysis/src/diagnostic/lua_diagnostic_code.rs
@@ -10,6 +10,8 @@ use serde::{Deserialize, Serialize};
 pub enum DiagnosticCode {
     /// Syntax error
     SyntaxError,
+    /// Lua syntax error
+    LuaSyntaxError,
     /// Type not found
     TypeNotFound,
     /// Missing return statement
@@ -66,6 +68,10 @@ pub enum DiagnosticCode {
     MissingReturnValue,
     /// Redundant return value
     RedundantReturnValue,
+    /// Undefined Doc Param
+    UndefinedDocParam,
+    /// Duplicate doc field
+    DuplicateDocField,
 
     #[serde(other)]
     None,
@@ -99,7 +105,7 @@ pub fn is_code_default_enable(code: &DiagnosticCode) -> bool {
     match code {
         DiagnosticCode::InjectFieldFail => false,
         DiagnosticCode::DisableGlobalDefine => false,
-        DiagnosticCode::UndefinedField => false,
+        // DiagnosticCode::UndefinedField => false,
         DiagnosticCode::IterVariableReassign => false,
         DiagnosticCode::CodeStyleCheck => false,
         // ... handle other variants