+++ crates/emmylua_code_analysis/src/diagnostic/test/missing_return_value_test.rs
@@ -16,7 +16,6 @@ mod tests {
         "#
         ));
 
-        
         assert!(ws.check_code_for(
             DiagnosticCode::MissingReturnValue,
             r#"
@@ -24,6 +23,22 @@ mod tests {
             ---@return string
             local function test()
                 return 1, "2"
+            end
+        "#
+        ));
+    }
+
+    #[test]
+    fn test_missing_return_value_variadic() {
+        let mut ws = VirtualWorkspace::new();
+
+        assert!(ws.check_code_for(
+            DiagnosticCode::MissingReturnValue,
+            r#"
+            --- @return integer?
+            --- @return integer?
+            function bar()
+                return
             end
         "#
         ));