+++ crates/emmylua_ls/src/handlers/completion/data/doc_tags.rs
@@ -1,4 +1,3 @@
-
 pub const DOC_TAGS: &[&str] = &[
     "class",
     "enum",
@@ -30,15 +29,3 @@ pub const DOC_TAGS: &[&str] = &[
     "source",
     "readonly",
 ];
-
-// // trivia tags just for completion
-// pub const DOC_OTHER_TAGS: &[&str] = &[
-//     "example",
-//     "note",
-//     "warning",
-//     "todo",
-//     "fixme",
-//     "author",
-//     "since",
-//     "throws",
-// ];